Rename DNP profile simulation module
