+++ dnplab/__init__.py
@@ -14,7 +14,7 @@ from .io.load import load
 
 from .analysis.relaxation_fit import *
 from .analysis.hydration import hydration
-from .analysis.enhancement_profiles import *
+from .analysis.simulate_enhancement_profiles import *
 
 from .processing import *
 from .widgets import *
