+++ dnplab/analysis/__init__.py
@@ -1,4 +1,4 @@
 """Modules which provide a workflow for analyzing dnpData objects"""
 from . import hydration
 from . import relaxation_fit
-from . import enhancement_profiles
+from . import simulate_enhancement_profiles
